remove arguably incorrect statement from Exercise 3.24 solution
a concise solution without pattern matching, using Scala's built-in
List type, is:

@annotation.tailrec
def hasSubsequence[A](l: List[A], sub: List[A]): Boolean =
  startsWith(l, sub) ||
    l.nonEmpty &&
      (startsWith(l.tail, sub) || hasSubsequence(l.tail, sub))
@annotation.tailrec
def startsWith[A](l: List[A], sub: List[A]): Boolean =
  sub.isEmpty ||
    l.nonEmpty && (l.head == sub.head) && startsWith(l.tail, sub.tail)

so though I may be nitpicking, I don't think it's really correct to
state that pattern matching saves the day here.  (in any case, the
statement isn't essential to include.)

the above code would need to be adjusted a little to work with the
book List instead of Scala's List, but the principle is the same.